Added linting (#3)
Added makefile directives `lint` and `reformat`, copying the settings from `fv3config`. Also added a circleci workflow which runs `make lint`.

Where I found unfinished/broken code that broke linting, I commented the code and added a line `raise NotImplementedError()`. On the lines where we want to add `[0, 0, 0]` to the left hand side but can't because of a gt4py bug, I added a TODO comment and a `# noqa` so the line will be ignored by the linter.

Commit history:

* Added make lint and make reformat, along with short description in README

* fix linting directives by adding python file list variables

* fix linting errors, ignore variable-unused errors

* fix unused variable linting errors

* fix linting errors on new files

* add linting tests to circleci

* enable only black linting for now

* add PHONY targets

* Added a bunch of missing PHONY targets