Add short-term quantity.storage method (#81)
Looking towards GDP-3 being implemented in gt4py, storages will be retrieved using `quantity.storage`, while the underlying `data` attribute will be kept as a numpy or cupy array. This PR implements an early version of that method which will work for our current test cases in fv3ser. Right now as a short-term solution, it is a shortcut to `.data`, until GDP-3 is implemented.