Add getters and setters for variables required by new dynamical core (#84)
This PR adds variables needed to call fv_dynamics in the new gt4py dynamical core. It does not yet include getters/setters for u_dt and v_dt, which require manual code as they do not live on the data container structures. u_dt and v_dt are only necessary when using the GFS_PHYS compiler flag (which we are).

Also:
- fixed a bug where atmosphere hybrid coordinates were incorrectly marked as cell center variables, and were missing one point
- removed some boilerplate test code which did not have any tests under it.