[VCMML-392] Add failing test for case with nans (#85)
The ZarrMonitor creates zarr arrays with a fill_value of 0 because `zarr.Group.create_dataset` has a default fill_value of 0 rather than None for some reason. Because xarray interprets any values equal to the fill_value as NaNs, this caused a very small, but meaningful number of NaNs to appear in the surface geopotential of the nudged run. See [VCMML-392].

This PR sets the fill_value to None which solves the problem. See the test (which was failing at first).

[VCMML-392]: https://vulcan.atlassian.net/browse/VCMML-392