Add quantity views and fill_scalar_corners routine (#2)
This PR adds regional views to Quantity for convenient indexing. `Quantity.view` now has attributes `northeast`, `northwest`, `southeast`, `southwest`, and `interior` which modify the indexing behavior of `view` to perform indexing relative to the edges and corners of the subtile domain.

We originally also planned to add `east`, `west`, `north`, and `south`, but I'm a little unsure about their APIs. It's unclear whether e.g. for `east`, the default slice for the Y-dimension should be the start to end of the data, or the start to end of the compute domain. Because the edge attributes are not needed for anything we have written so far in fv3core, I have left them out in this PR. They are easy to add in (the syntax is the same as for the corners). I am including the tests I implemented for these as comments so they are committed to version control - they could be removed in a subsequent PR.

I've also added HISTORY file, and moved fv3util-relevant entries since the last release.

Because these changes are copied in from the fv3gfs-python repo, there may be some errors like duplicated files or changes being reverted which I did not want to revert - I've removed a couple (hopefully all) of these, but please comment if you spot any more.