Use tags when performing halo updates (#13)
We were getting indefinite hangs sometimes when performing certain sequences of sends and receives, due to sends not being paired with the correct receives. This PR uses tagged send/recvs on each halo update. The n'th halo update is tagged with the integer n-1.

No changes are made to public API.