remove namelist values as runtime args in saturation_adjustment (#105)
Gt4py has a bug that does not allow runtime constants to be used as a conditional, and we had a workaround in saturation adjustment that involved hard coding namelist values as constant. But now that we import the entire namelist as an external variable (SimpleNamespaces are supported), we can use the namelist values directly in the stencils without passing them as arguments.