use MPI terminology to refer to root rank (#24)
Currently the partitioner and communicator modules refer to the MPI root rank as the master rank. This PR changes variable names and documentation to be consistent with the MPI terminology. While this is technically a backwards incompatible change, I do not expect it to break external code as the symbols appear only to be used internally. The refactored symbols are not used in fv3net.