fix bug where quantity.view could become out of sync with data (#27)
Accessing `.storage` on a quantity which was initialized with a numpy or cupy array and a gt4py_backend value currently causes `view` to become out of sync with the underlying data of the quantity, since the data is re-allocated in creating the storage but the view is unchanged and retains a reference to the old data. This PR updates the routine which re-allocates the data to also update the view.

This is a work-around to deal with not being able to initialize gt4py storages from existing numpy or cupy arrays. We should be able to remove the work-around after GDP-3 (https://github.com/GridTools/gt4py/pull/28) is merged.