Implement scatter/gather for CubedSphereCommunicator (#33)
In preparing examples for publication, it will be helpful to plot global variables. This PR adds a cube gather routine to enable this. Scatter is added as well for completeness and other use cases, and because it was simple to do so. Most of the implementation is a re-use of the code for tile scatter and gather.

The current implementation requires (and enforces with NotImplementedError) that the tile dimension be the first dimension of the global Quantity. This may or may not be relaxed in the future. The rest of the dimension order is unrestricted.

Changes:
- Added TILE_DIM constant for tile dimension of global quantities
- Added Partitioner base class implementing features necessary for scatter/gather
- Moved scatter and gather from TileCommunicator to the Communicator base class, so its code can be re-used by the CubedSphereCommunicator
- implemented subtile_slice, global_extent, and subtile_extent routines on CubedSpherePartitioner necessary for scatter/gather in CubedSphereCommunicator.
- renamed argument `tile_extent` and `tile_dims` to `global_extent` and `global_dims` in routines to refer generically to the tile in the case of tile scatter/gather or cube in the case of cube scatter/gather

Tests are added for scatter/gather on the cube. While direct tests are not included for CubeSpherePartitioner's subtile_slice, global_extent, or subtile_extent routines, these are tested by the scatter/gather tests, which only assume TilePartitioner's versions of these routines are correct.