Fix setup.py and test it with tox in Circle CI (#50)
Currently the setup.py file is missing some dependencies and refers to files not in `MANIFEST.in` which sometimes prevent this package from being `pip installed`.  This will make it hard to distribute this package via PyPi. These errors appeared because the code in`setup.py` is not actually being tested.

This PR adds the missing requirements (scipy and mpi4py) to the setup.py and tests it with [tox](https://tox.readthedocs.io/en/latest/). Tox is a widely used tool for testing python projects in isolated virtual environments and ensuring that the `setup.py` actually works. Another advantage is that it moves testing code out of the circle/config.yml and allows the local developer to trigger the tests identically to the CI.

~~The build time increased because tox now runs many thousands of more tests than before.~~

Other changes:
- removed caching logic from CircleCI. This caching logic would have to change for tox, and the environment install only takes about 1 minute, so it's probably not worth the complexity to optimize this portion of the build. This probably changed when `numcodecs` started uploading wheels.

Edit: I fixed this in 031eee4