Memoize outputs of make_storage_from_shape (#290)
## Purpose

Currently we re-allocate storages in many places within loops. This PR adds memoization to `make_storage_from_shape`, so that the same storage can be provided each time the same storage is requested. There are some caveats and gotchas to this implementation, and in the long run we should replace it with a safer implementation.

The key used for memoization is the arguments used combined with the calling scope file and line number, as well as the file and line number which called in to that scope. This handles cases where a utility function (such as `copy`) calls our `make_storage_from_shape`, since `copy` will be called from different places each time. This does *not* handle any more deeply nested duplicate calls, such as if another utility function were to call `copy`, and does not handle allocations which take place within for loops, such as tracer allocations. In those cases, memoization will provide the same storage to two conceptually different objects, causing a bug. For this reason, and because of the significant overhead cost of `inspect`, we should move away from this implementation in the longer term.

In the short term, this reduces our time spent allocating storages on each timestep in our profiling job from 130s to 5s, plus an inspect overhead of about 35s. When we move to another implementation, this inspect overhead can also be removed.

To avoid having to zero out the storages each time a memoized storage is returned, the default behavior for `make_storage_from_shape` has been changed to return `empty` arrays. In places where it is critical to initialize zeroed arrays, the call has been changed to explicitly set `init=True`.

## Code changes:

- Memoization has been added to `get_storage_from_shape`, keyed on the arguments as well as the filename and line number of the two frames above the current frame (the calling filename and line number, and also the filename and line number which called into that scope).
- Default value for `init` in `make_storage_from_shape` has been changed to `False`, and code which requires zero values has been modified to explicitly set `init=True`