enable mypy in pre-commit, fix simplest typing errors (#360)
## Purpose

While we have type hints in our repo, they are currently untested and in some cases misleading. This PR enables static type checking through mypy to ensure they reflect the code written, and to provide more robustness in the code (by testing typing, which we don't currently do in our unit tests). This can give faster turnaround on feedback when refactoring the codebase than the unit tests.

In this commit type checking is disabled for many files in the repository. They can be progressively enabled by deleting the corresponding ignore line from `pre-commit-config.yaml`.

Gt4py breaks static typing when for some operations, requiring an ignore comment. For example:
```
with interval(0, 1):
    peg = ptop
with interval(1, None):
    peg = peg[0, 0, -1] + 1.  # type: ignore[index]
```

This shows up in only a few places in our repo.

## Code changes:

- Fixed the most straightforward typing errors, particularly for files with very few errors
- `gtstencil()` now only operates as a function which returns a decorator. It was being used this way in almost all cases, and the remaining cases could be modified to use it this way by only adding parentheses. This greatly simplifies type checking for this routine, and makes its usage uniform across the repository.

## Requirements changes:

- Added mypy to `pre-commit-config.yaml`

## Infrastructure changes:

- Added mypy type checking for certain files in the repo