Optimize origin and domain handling (#369)
Purpose

Our performance investigations showed that the calls to normalize the origin for each stencil are unnecessarily expensive. This PR solves that problem by moving the normalization process to the initialization-time of stencil objects. Additionally an lru_cache is applied to grid.axis_offsets to further reduce the expensive calls to the gt4py_tuple

Code changes:

    Addition of the FixedOriginStencil that requires origin and domain at instantiation-time and stores the normalized origins
    Example usage of the FixedOriginStencil in FVTP2D
    Addtion of LRU caching to grid.axis_offsets

Infrastructure changes:

    Switch to the GT4Py version that offers the functionality to pass in normalized origins