perform tracer halo updates in parallel (#390)
## Purpose

Currently tracer halo updates in tracer_2d_1l occur one at a time, blocking between tracers. This PR refactors that logic to start all tracer halo updates, and then only wait on the halo update for a given tracer when it is being worked on. This should be more efficient if it results in more synchronous communication.

## Code changes:

- tracer halo updates in tracer_2d_1l now occur asynchronously