make xarray and zarr optional dependencies, test with and without them (#66)
In testing fv3core on HPC, we are constrained in the number of files we can put on the filesystem. Most of these files are python files in our runtime environment. Making xarray and zarr optional dependencies will allow us to test fv3core on HPC without them, reducing our file count.