Automatic Device Synchronization (#424)
This PR enables fv3core to use automatically synchronized GPU storages in gt4py. That code is found in PR #409 and also implemented in gt4py-develop tag v26. This allows all of the manual device synchronizations to be removed from the dynamical core code. Note that the device syncs in the test translate classes have not been removed in this PR.

Code changes:
- Remove StencilWrapper.disable_cache option as unused.
- Add StencilWrapper.run method to encapsulate _process_kwargs and stencil_object.run in one call.
- Add StencilWrapper.write_fields property so the run method can call _set_device_modified on the READ_WRITE fields.
- Remove utils.device_sync() calls from the dynamical core code.

Requirements changes:
- Update GT4PY_VERSION to v26.