OOP-ify Fast Corners (#452)
## Purpose

Copy Corners showed up as one of the slowest stencils we run. This was due to it creating a kernel call for each region executed. This PR speeds up the corner computation. 
As this functionality  is used in three spots throughout the code, we create an object to re-use the logic.


## Code changes:

- Create a `CopyCorners` class that handles the logic of copying to the temporary in the separate kernel call
- Applies the object in all two spots (ignoring delnflux as is is currently in the process of being refactored)