Simplify stencil wrapper classes and caching logic (#442)
## Purpose

This PR simplifies the classes used for wrapping stencils, removing features we no longer use and adding tests for the features we do use.

It also adds a requirement that stencil origins must be explicitly defined, for example if an input has a singleton dimension its origin must be zero for that dimension. Code which relied on gt4py having a side-effect of changing this origin to `shape[i] - 1` has been updated to explicitly pass the origin, or no longer use a singleton-dimension array. In other words, when we need the origin to be different for different inputs, we now explicitly pass those origins.

## Code changes:

- Added `StencilConfig` class which contains all configuration used to initialize a gt4py stencil
- stencils now get all global gt4py configuration from a new method global_config.get_stencil_config(), instead of a separate getter for each configuration option
- "stencil report" functionality which was used for initial validation but has not been used in months is removed
- when using gtstencil as a decorator, the syntax is now `@gtstencil` with no parentheses. To decorate with explicit externals, a separate assignment line must be used.
- The error produced when trying to use a stencil which has fields come after parameters is now more informative