validate zh in UpdateDzD on only compute domain, remove temporary (#389)
## Purpose

Currently our validation tests check that the output arrays are identical to the Fortran outputs, including in halo regions. For some stencils this is valid because halo values are updated and then used by other stencils, but in many cases it is only necessary to test a subset of the output for validation, because the halo values are overwritten without being read.

This PR adds a solution to validate only a subset of the output, while also guaranteeing the non-validated data is unused by setting it to NaN during validation tests (including larger tests like fv_dynamics and dyn_core). In this PR that solution is used to validate zh in UpdateDzD on only the compute domain, allowing us to remove a temporary copy of zh whose compute domain is never modified.

## Code changes:

- Validate zh on only the compute domain in UpdateDeltaZOnDGrid, setting halo output values to nan during tests
- Remove z2/zh_tmp temporary from UpdateDeltaZOnDGrid, using zh directly instead
- Added SelectiveValidation class which provides a slice to subset arrays for validation, a method to set arrays to NaN when the class is in test mode, and a dictionary for translate class output specifications
- Updated test_translate to check translate classes for a `subset_output(varname, array) -> array` method which subsets outputs for validation (or returns the unmodified array). If the method is not present, the behavior is unchanged.

## Infrastructure changes:

- parallel test command is updated so that they fail when any worker fails, instead of deadlocking if only a subset of workers fail