when jenkins runs, even on master, it checks out a commit and thus using git branch to get whether master is the name of the branch or not does work. jenkins brovides GIT_BRANCH, which I think we want anyways, because it would prevent saving artifacts if run on master by hand (#468)
A recent commit stopped the copying of performance results to the location picked up by the performance plots. When jenkins runs, even on master, it checks out a commit and thus using git branch to get whether master is the name of the branch or not does work. jenkins provides GIT_BRANCH, which I think we want anyways, because it would prevent saving artifacts if run on master by hand