make pfull init time (#462)
Purpose
Remapping computes a k origin of kmp using pfull in the runtime. This could be resolved by pushing the pfull comparison as a runtime if condition inside of stencils, but this will not look great. pfull actually only depends on things known at the init time of the FVDynamics objects that don't change, so it also can be computed in init. Then when we make remapping an object, we can make the computation of kmp also init time, and use FrozenStencil for stencils that use kmp as an origin.

Code changes:
Updated fv_dynamics and dyn_core to use init time pfull.