add an assertion to prevent a stencil from being double decorated (#482)
To prevent a likely mistake that only gets caught in the standalone as we refactor. If we accidentally try to decorate a stencil that has already been decorated (e.g. forgot to remove @gtstencil before the definition). This assertion should also just make sense to have, a stencil with no arguments doesn't make any sense.