Reintroduce while loop version of vertical remapping (#483)
This PR reintroduces the while-loop implementation of the Lagrangian-Eulerian remapping algorithm. The code was reverted in the belief that it was causing illegal memory access errors in CUDA. Those errors have since been identified to be caused by the lack of device syncs in the compilation step of the run_on_daint script because it was disabling halo exchanges. The recent introduction of new halo exchange code has also solved the problem of compilation steps crashing when run in a distributed context. Hence, the --disable_halo_exchange flag is removed. Finally, the GT4PY_VERSION is updated to v29 which includes a bug fix for the auto-synchronizing storages.

Code changes:
- Reintroduce while loop implementation of the vertical remapping algorithm.
Requirements changes:
- Update the gt4py version to v29.
Infrastructure changes:
- Remove the --disable_exchange_flag from the run_on_daint.sh script.