Surface vertical velocity is always 2d (#494)
Purpose
There was a conversion for wsd between 2d and 3d every timestep as parts of the code optimized to use it for the 2d nature it really is, and it was much more involved to unravel the 3d nature from this variable in remapping at the time. This PR unravels that, so we can always use wsd in 2d. The time converting is minimal, but as we investigate GPU profiles, we are seeing it as CPU work chunks that don't need to happen anymore.

Code changes:
removed CPU assignment statements for wsd_3d and wsd in fv_dynamics
wsd and the qs variable is becomes at lower levels is converted to a FloatIJ
Since remapping uses wsd at the deepest level, remap_profile, and sometimes a different variable, gz, which is not 2d was passed to remap_profile, this touches many of the remapping files.
In the fortran code 'gz' is passed to map1_ppm for iv modes where qs is not actually used, so has no impact. Instead define a temporary variable in the MapSingle class that can be used when an actual variable is not passed it. gz is no longer passed in, it was never used.
test code is updated to make wsd 2d, but there may be an option to do this more generally rather than duplicating the logic. I am open to putting more effort here to make that cleaner.