Oop/divdamp (#475)
Purpose
Make a class DivergenceDamping to pull out more unnecessary runtime operations into an init phase. We have a separate task to turn these many stencils into less stencils using regions, and that is not part of this. It should thankfully remove most of the complexity presented here.
Also remove the loop over k subsets, making all the stencils in here 3d (well there are two sets of stencils, ones that go from 0 to nonzero_nord_k, and those that go from that to the top). Because nord is 0 for some number of low k levels, then it is namelist.nord for the rest of the column, there are different operations on 2 ranges.

Code changes:
divergence_damping is restructured to be a class
a loop over nord is computed, here we initialize stencils in init time for each pass of this loop.
the call to divergence damping is done once, not in a k-loop
FillCornersBGrid class added imitating the structure of CopyCorners to stop the excessive kernel calls q=q causes
A2B_Ord4's test call to divergence_damping.vorticity_calc is updated to grab that method from a DivergenceDamping object