fvsubgridz is 3d and OOPed (#511)
Purpose
fv_subgridz is outside of the main 'fv_dynamics' loop and often gets left behind in refactors, and is the last piece preventing us from completing the transition from stencils decorated in the file and compiled JIT to one using objects to initialize/compile the whole model before running it. This PR catches it up, though a choice was made to unroll all of the tracers operations, with the idea we'll be adding a gt4py feature that can do this loop inside of the stencil. It was necessary to do so for the m_loop to be consolidated, and changed the other stencils to make them consistent.
The main purpose of this PR:
-- make a single stencil inside of the n-loop rather than looping over k to adjust both this layer and the one below it simultaneously.
-- make FVSubgridZ an object that matches the conventions of our other stencils.
What this does not do:
-- m_loop has a lot of repetition. We could make more effort to use more functions, but most variations tried resulted in failed validation. More could be done here, but in favor of getting this code up to date I suggest we delay that to a future optimization, when we revisit it to apply the loops over tracers.

Code changes:
fv_subgridz is rewritten to be 3d and as a class
removed remaining references to 'sd' and 'si'
updated example runfiles to call the FvSubgridZ object