cleanup gtstencils (#512)
Purpose
We transitioned from using decorated @gtstencil s to initializing classes and freezing stencils. Although the gtstencil method provides a mechanism for freezing any unfrozen stencils in the first model timestep, to simplify the conventions the model is using, we've pushed to get everything frozen in the same way. This PR does a final cleanup of leftover @gtstencils, freezing them or removing them when not needed. Also removed the last case of a stencil class being cached. I left the gtstencil mechanism in place in case we find a future need for it (e.g. decide to shift back to more JIT rather than precompiling everything), but I am open to removing it to make our model code as simple and easy to follow as we can.

Code changes:
updatedzc was still using a @gtstencil in corners, changed that to be frozen
removed fill_corners_cells, which was a neat abstraction, but no longer used by the model besides in a test, so updated the test and removed it
fv_dynamics gets a frozen copy stencil
removed the fv_dynamics.fv_dynamics method and updated example runfiles to initialize a DynamicalCore instance