Getting rid of the copy_j_adjacent stencil (#529)
Purpose
The copy_j_adjacent stencil breaks the gtc backends and is a small (and 2D) stencil to have to call separately. This gets rid of it and pulls that functionality into update_ua

Code changes:
remapping.py: removes the copy_j_adjacent stencil. Instead this uses a region in update_ua to correctly assign values to ua, though we could also use a region in moist_cv_pt_pressure to set pe2 at je+1