GTC regions safety/a2b ord4 (#553)
Purpose
A2B_Ord4 has regions that depend on computations in neighboring regions. This can be a race condition and also is not efficient. Previous attempts to inline these same one resulted in memory issues, but that was with a more aggressive refactor of A2B_Ord4 to reduce the stencil count. Here we do the minimal necessary to unblock testing fv3core with regions in GTC.
It has been verified that this change does not slow does the code (https://jenkins.ginko.ch/job/fv3core-performance-test/573/backend=gtcuda,experiment=c128_6ranks_baroclinic,slave=daint_submit/artifact/2021-07-07-06-30-26.json)

Code changes:
a2b_ord4 has a light refactor to inline region calculations
Infrastructure changes:
Update the DEV=y setting with the Makefile to not docker build when that was set, as was intended with the previous dev_tests endpoint