fv3core works with GTC without regions (#560)
Purpose
In the feature/gtc-dace branch, we removed regions from fv3core and tested the rest of the code. This PR brings the changes needed we found there once FVDynamics validated with the gtc:gt:cpu_ifirst and gtc:gt:gpu backwnd, as well as some optimizations done along the way. With this, and the known regions race conditions we are also fixing, we should be able to test fv3core with regions with the GTC backends.

NOTE, there is one change needed in D_SW:heat_source_from_vorticity_damping for the gtc:gt:gpu to validate that breaks the old (current) gtcuda backend. So instead of a code change a TODO was added. When we shift to GTC, that code change will need to happen at the same time.

Code changes:
Fix to D_SW calculation of horizontal relative vorticity by adding a computation, otherwise reading with an offset API fields that have just been written does not work.
DivergenceDamping won't run the stencils for 'low k values' where nord=0, if there are in fact 0 vertical indices that meet this criteria. Otherwise some backends crashed.
NegAdj3 no longer has overlapping intervals and also has terniary conditions that involve self assignment turned into simple if statements
RemapProfile has several refactors. On changes a PARALLEL to a FORWARD to prevent out of order calculations(workaround for a potential bug), factoring terniary if statements with self assignment to simple if statments and logic condensation to make it more efficient and other requirements to pass on GTC (@eddie-c-davis, would you mind updating this if there is anything else ?). Also the 'ext' variables are BoolFields -- not required for GTC but was a concern setting a float to a logical value and checkin == 1 could be a recipe for trouble.
SaturationAdjustment uses hydrostatic INLINED and temporary fields are cleaned up -- not required for GTC, but changes came about when investigating failures, and looks to be more efficient.
Remove an overlapping interval from temperature_adjust
Overrides file includes gtc:gt:gpu and gtc:gtcuda backends
Infrastructure changes:
added installation of the GridTools v2 to the jenkins install venv script. Instead this should be done in daint_venv when we are ready to move to GTC (or anytime before)
added (Eddie) a global config method to check wither the backend is gpu rather than matching the cuda string