Add GridData, DampingCoefficients classes (#559)
Classes are used in currently-refactored inits. d_sw is partially refactored, there's a bit more work to remove the old Grid class from it entirely but I'd like to push the changes so far through and complete d_sw in a follow-on PR.

## Purpose

We looked at what grid variables are generated by the grid savepoints and where those grid variables are used to determine some container classes. Variables that are used in only one place will later be calculated in that init routine, for now I've converted those to access global grid so they stay out of the init signatures. These should be cleaned up when the grid generation code is available.

Some grid variables not used by the dynamical core but used by the physics are included as properties that raise NotImplementedError. This includes `lat`, `lon`, `ptop`, `ak`, and `bk`.

## Code changes:

- Created `GridData` class composed of `ContravariantGridData`, `AngleGridData`, `HorizontalGridData`, and `VerticalGridData`
- Created `DampingCoefficients` class
- Defined properties on `Grid` to retrieve these new classes
- Refactored DivergenceDamping, AGrid2BGridFourthOrder, DGrid2AGrid2CGridVectors, DelnFlux, DelnFluxNoSG, FiniteVolumeFluxPrep, XTP_U, YTP_V, and partially DGridShallowWaterLagrangianDynamics to use new data container classes