Create config classes for d_sw, riemann solver, and AcousticDynamics (#573)
## Purpose

The namelist currently exists only as a flat namespace, with individual parameters being passed to many internal routines to handle data containerization. This can result in many arguments (when data is not containerized), or being unable to determine what namelist parameters are being used (when it is containerized).

This PR creates configuration container classes for AcousticDynamics, DGridShallowWaterLagrangianDynamics, and RiemannSolver3 which contain only the data they require. These can be retrieved as attributes from the top-level namelist.

## Code changes:

- Added classes AcousticDynamicsConfig, DGridShallowWaterLagrangianDynamicsConfig, and RiemannConfig
- Added attributes `riemann`, `d_grid_shallow_water`, and `acoustic_dynamics` to `Namelist`